Refuse to land an inactive revision unless forced.
Summary:
If diff used to land revision is not the active one Lando API will
fail with a 409 error. Client will then provide an information to the user
about the issue and that it is possible to force it.
In such case client will request a new landing with a
force_override_of_diff_id parameter equal to the active id. API will proceed
with the landing.

Active diff id is saved in database in case it would change in the
future.

Two additional fixes:
- Fixed creation of phid response in `tests.factories`
- Renamed the `land` method to `post` in `api.landings`
- Removed deoendency of `json` in `api.landings`
- Added a msg to the RevisionNotFoundException
- Removed space at the end of the line in swagger

Test Plan: invoke test,
invoke upgrade before testing in UI

Reviewers:

Subscribers:

Bug #: 1396602