Add Postgres support
Dockerfile needs to install gcc and musl-dev to be able to compile
psycopg2.
Status column is migrated from Integer to String. There was no exception
thrown before and sqlite was saving String to his column.
Migration is done to upgrade only. SQLite allows to write a string to
an integer column type. We should never downgrade.
Result has a default value of empty string.

Tests are still running using SQLite.