WIP: Check diff for integrity
User is choosing a diff and revision to be landed. It is not checked if
the diff is actually related to the requested revision. This patch fixes
this issue.
New method is added `check_integrity` which is called from `Patch.build`
method. It loads the diff from Phabricator and checks if the
`revisionID` equals the revision id of the requested revision.