Dev setup: Using PostgreSQL on dev machines
Summary:
We're using SQLite on dev machines for testing and running the Swagger
UI. Our developer setup should be as close to production as possible.
This patch is providing PostgreSQL for developer machines.

A PostgreSQL container is created. It runs 2 databases - one for tests
and another for developer server (Swagger UI).

Refactored:
Readme.md - headings, contents of the running the development server.

Reviewers: imadueme, dylan

Bug #: 1417587

Differential Revision: https://phabricator.services.mozilla.com/D249