Landing: Storing revision id to integer.
Our system is inconsistent about format of the `revision_id`. Sometimes
it is used as an integer and another as a string. This patch is changing
the behaviour to use integer inside the application. On the "outside"
`revision_id` is presented as a string in the format of `'D{number}'`.
Value is translated from string to integer on the input and translated
back to the string on the output.