Landing - Do not land if any parent is open.
We do allow to land the revision with an open dependency. We will now stop the
landing process if an open dependency has been found.

This patch is recursively checking the dependency tree for any open
parent. If any is found API is responding with a 409 HTTP Error. It
includes a comment with the id of the open dependency,